Switch to assigning /24s out of 169.254/16 instead of 10.100/16.
10/8 is heavily used in private networks, making conflicts more
likely, while 169.254/16 is specifically reserved for link-local
addresses. Since debmarshal networks are always isolated, these
addresses will always be link-local.

This is sort of in conflict with RFC 3927, which governs assignment
and usage of 169.254/16, but since we control all interfaces on
debmarshal networks, it doesn't really seem like that big of a deal.

Signed-off-by: Evan Broder <ebroder@google.com>
Signed-off-by: Drake Diedrich <dld@google.com>

git-svn-id: https://debmarshal.googlecode.com/svn/trunk@203 485416fb-281f-0410-a025-fb7d5554b9fe