Use setuptools entry_points instead of a "classId" to find Distribution subclasses.
This will make it easier to find out-of-tree classes which implement
support for a new distribution. setuptools is also a more standardized
way for implementing application extensibility.

Note that this won't be a security concern, because the entry_points
will be loaded by the privileged daemon, which won't be affected by
any non-globally-installed modules and eggs.

Signed-off-by: Evan Broder <ebroder@google.com>

git-svn-id: https://debmarshal.googlecode.com/svn/trunk@321 485416fb-281f-0410-a025-fb7d5554b9fe