Everybody loves locking! Add some to the disk image generation code.
The shared lock on the base image when generating a custom image looks
questionable, but it should actually be safe. While this does allow
the image to be mounted multiple times, it shouldn't be written to by
userspace, and should only be written to by kernelspace to replay the
journal, which should only happen the first time it's mounted (making
subsequent mounts safe).

Signed-off-by: Evan Broder <ebroder@google.com>
Signed-off-by: Drake Diedrich <dld@google.com>

git-svn-id: https://debmarshal.googlecode.com/svn/trunk@325 485416fb-281f-0410-a025-fb7d5554b9fe